refactor(api): reorganize api tests in contexts
After changing production code we follow the API contexts structure in
tests too.
