+++ tests/api/tests/context/mod.rs
@@ -1,0 +1,4 @@
+pub mod auth_key;
+pub mod stats;
+pub mod torrent;
+pub mod whitelist;
