+++ tests/api/tests/mod.rs
@@ -1,0 +1,4 @@
+pub mod authentication;
+pub mod configuration;
+pub mod context;
+pub mod fixtures;
